requirements.txt
diff-match-patch==20121119
Django==1.11
django-ckeditor==5.5.0
django-import-export==1.0.1
django-js-asset==1.1.0
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pytz==2018.4
PyYAML==4.1
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
